Add support for previews of links to images
This adds support for displaying an overlay window when hovering over a
link to an image.

This uses cURL for fetching the image, and decodes and rescales it in a
streaming manner with GdkPixbufLoader. Moving the cursor way from the
link or from its starting position closes the window.

Only links ending with .jpg and .png are supported for now.

It seems that GTK doesn't send multiple events as a rescaled image is
processed, but only a single event when it is done, which prevents most
large images from being displayed progressively as they are downloaded.